Add a section on events the Society runs
The section formalises the fact that attendees of the events that
the Society runs must abide by a code of conduct.
https://hackcodeofconduct.org/

It also adds a set of model rules for game jams the Society runs,
whilst stressing that the organiser of individual events may vary them
as they see fit.